server-PC/requirements.txt
Flask==1.0.2
pyserial==3.4